feat: adiciona validação no campo de entrada de investimento inicial
